Add a little bit of TDR handling
It's not exactly possible for Gelly to handle TDR issues as FleX has no known documentation on this. However, the cause is now able to be logged during a TDR event.

+++ packages/gelly-gmod/src/main.cpp
@@ -477,8 +477,9 @@ LUA_FUNCTION(gelly_SetRenderSettings) {
 	GET_LUA_TABLE_MEMBER(int, SmoothingIterations);
 	GET_LUA_TABLE_MEMBER(int, ThicknessIterations);
 
-	newSettings.filterIterations = SmoothingIterations;
-	newSettings.thicknessFilterIterations = ThicknessIterations;
+	newSettings.filterIterations = static_cast<int>(SmoothingIterations);
+	newSettings.thicknessFilterIterations =
+		static_cast<int>(ThicknessIterations);
 
 	gelly->SetRenderSettings(newSettings);
 	CATCH_GELLY_EXCEPTIONS();
